Conditionals and trimming
Deleted all extra states and events, renamed one of the unready events to unready event. Added Boolean fields to the Context for the conditional evaluation.
